+++ CHANGELOG.md
@@ -8,10 +8,14 @@ and this project (partially) adheres to [Semantic Versioning](https://semver.org
 ## [Unreleased]
 ### Added
 - `Model` support for Raspberry Pi 5+
+- `none` logo style to completely hide distribution logo
 - AppArmor confinement profile (included in Debian and AUR packages)
 
 ### Changed
 - `Model` honor `/proc/device-tree/model` when it exists
+
+### Removed
+- `_distribution` protected attribute from `Output` class
 
 ## [v4.14.3.0] - 2024-04-06
 ### Added